HSDIOError circular import fix
