+++ ni_hsdio.py
@@ -4,25 +4,7 @@ import struct
 import platform # for checking the os bitness
 import logging
 from typing import Tuple
-from hsdio import HSDIO
-from pxierrors import HardwareError
-
-
-class HSDIOError(HardwareError):
-    """
-    Raised for errors coming from NI HSDIO drivers
-
-    Attributes:
-        error_code : Integer code representing the error state
-        message : message corresponding to the error_code with some traceback info
-    """
-    def __init__(self, error_code, message):
-        self.error_code = error_code
-        super().__init__(
-            device=HSDIO,
-            task=HSDIOSession,
-            message=message
-        )
+from pxierrors import HSDIOError
 
 
 class HSDIOSession:
